+++ crates/sncast/tests/e2e/account/import.rs
@@ -25,11 +25,11 @@ pub async fn test_happy_case(input_account_type: &str, saved_type: &str) {
         "--accounts-file",
         accounts_file,
         "account",
-        "add",
+        "import",
         "--url",
         URL,
         "--name",
-        "my_account_add",
+        "my_account_import",
         "--address",
         "0x123",
         "--private-key",
@@ -43,7 +43,7 @@ pub async fn test_happy_case(input_account_type: &str, saved_type: &str) {
     let snapbox = runner(&args).current_dir(tempdir.path());
 
     snapbox.assert().stdout_matches(indoc! {r"
-        command: account add
+        command: account import
         add_profile: --add-profile flag was not set. No profile added to snfoundry.toml
     "});
 
@@ -55,7 +55,7 @@ pub async fn test_happy_case(input_account_type: &str, saved_type: &str) {
         json!(
             {
                 "alpha-sepolia": {
-                  "my_account_add": {
+                  "my_account_import": {
                     "address": "0x123",
                     "class_hash": DEVNET_OZ_CLASS_HASH_CAIRO_0,
                     "deployed": false,
@@ -79,11 +79,11 @@ pub async fn test_existent_account_address() {
         "--accounts-file",
         accounts_file,
         "account",
-        "add",
+        "import",
         "--url",
         URL,
         "--name",
-        "my_account_add",
+        "my_account_import",
         "--address",
         DEVNET_PREDEPLOYED_ACCOUNT_ADDRESS,
         "--private-key",
@@ -102,7 +102,7 @@ pub async fn test_existent_account_address() {
         json!(
             {
                 "alpha-sepolia": {
-                  "my_account_add": {
+                  "my_account_import": {
                     "address": DEVNET_PREDEPLOYED_ACCOUNT_ADDRESS,
                     "class_hash": &DEVNET_OZ_CLASS_HASH_CAIRO_1.into_hex_string(),
                     "deployed": true,
@@ -126,11 +126,11 @@ pub async fn test_existent_account_address_and_incorrect_class_hash() {
         "--accounts-file",
         accounts_file,
         "account",
-        "add",
+        "import",
         "--url",
         URL,
         "--name",
-        "my_account_add",
+        "my_account_import",
         "--address",
         DEVNET_PREDEPLOYED_ACCOUNT_ADDRESS,
         "--private-key",
@@ -144,8 +144,8 @@ pub async fn test_existent_account_address_and_incorrect_class_hash() {
     let snapbox = runner(&args).current_dir(tempdir.path());
 
     snapbox.assert().stderr_matches(formatdoc! {r"
-        command: account add
-        error: Incorrect class hash {} for account address {}
+        command: account import
+        error: Incorrect class hash {} for account address {} was provided
     ", DEVNET_OZ_CLASS_HASH_CAIRO_0, DEVNET_PREDEPLOYED_ACCOUNT_ADDRESS});
 }
 
@@ -158,11 +158,11 @@ pub async fn test_nonexistent_account_address_and_nonexistent_class_hash() {
         "--accounts-file",
         accounts_file,
         "account",
-        "add",
+        "import",
         "--url",
         URL,
         "--name",
-        "my_account_add",
+        "my_account_import",
         "--address",
         "0x202",
         "--private-key",
@@ -176,7 +176,7 @@ pub async fn test_nonexistent_account_address_and_nonexistent_class_hash() {
     let snapbox = runner(&args).current_dir(tempdir.path());
 
     snapbox.assert().stderr_matches(indoc! {r"
-        command: account add
+        command: account import
         error: Class with hash 0x101 is not declared, try using --class-hash with a hash of the declared class
     "});
 }
@@ -190,11 +190,11 @@ pub async fn test_nonexistent_account_address() {
         "--accounts-file",
         accounts_file,
         "account",
-        "add",
+        "import",
         "--url",
         URL,
         "--name",
-        "my_account_add",
+        "my_account_import",
         "--address",
         "0x123",
         "--private-key",
@@ -206,8 +206,8 @@ pub async fn test_nonexistent_account_address() {
     let snapbox = runner(&args).current_dir(tempdir.path());
 
     snapbox.assert().stderr_matches(indoc! {r"
-        command: account add
-        error: There is no contract at the specified address
+        command: account import
+        error: Class hash for the account address 0x123 could not be found. Please provide the class hash
     "});
 }
 
@@ -220,32 +220,28 @@ pub async fn test_happy_case_add_profile() {
         "--accounts-file",
         accounts_file,
         "account",
-        "add",
+        "import",
         "--url",
         URL,
         "--name",
-        "my_account_add",
+        "my_account_import",
         "--address",
         "0x1",
         "--private-key",
         "0x2",
-        "--public-key",
-        "0x759ca09377679ecd535a81e83039658bf40959283187c654c5416f439403cf5",
-        "--salt",
-        "0x3",
         "--class-hash",
         DEVNET_OZ_CLASS_HASH_CAIRO_0,
         "--type",
         "oz",
         "--add-profile",
-        "my_account_add",
+        "my_account_import",
     ];
 
     let snapbox = runner(&args).current_dir(tempdir.path());
 
     snapbox.assert().stdout_matches(indoc! {r"
-        command: account add
-        add_profile: Profile my_account_add successfully added to snfoundry.toml
+        command: account import
+        add_profile: Profile my_account_import successfully added to snfoundry.toml
     "});
     let current_dir_utf8 = Utf8PathBuf::try_from(tempdir.path().to_path_buf()).unwrap();
 
@@ -257,13 +253,12 @@ pub async fn test_happy_case_add_profile() {
         json!(
             {
                 "alpha-sepolia": {
-                  "my_account_add": {
+                  "my_account_import": {
                     "address": "0x1",
                     "class_hash": DEVNET_OZ_CLASS_HASH_CAIRO_0,
                     "deployed": false,
                     "private_key": "0x2",
                     "public_key": "0x759ca09377679ecd535a81e83039658bf40959283187c654c5416f439403cf5",
-                    "salt": "0x3",
                     "legacy": true,
                     "type": "open_zeppelin"
                   }
@@ -274,8 +269,8 @@ pub async fn test_happy_case_add_profile() {
 
     let contents = fs::read_to_string(current_dir_utf8.join("snfoundry.toml"))
         .expect("Unable to read snfoundry.toml");
-    assert!(contents.contains("[sncast.my_account_add]"));
-    assert!(contents.contains("account = \"my_account_add\""));
+    assert!(contents.contains("[sncast.my_account_import]"));
+    assert!(contents.contains("account = \"my_account_import\""));
     assert!(contents.contains(&format!("url = \"{URL}\"")));
 }
 
@@ -288,11 +283,11 @@ pub async fn test_detect_deployed() {
         "--accounts-file",
         accounts_file,
         "account",
-        "add",
+        "import",
         "--url",
         URL,
         "--name",
-        "my_account_add",
+        "my_account_import",
         "--address",
         DEVNET_PREDEPLOYED_ACCOUNT_ADDRESS,
         "--private-key",
@@ -304,7 +299,7 @@ pub async fn test_detect_deployed() {
     let snapbox = runner(&args).current_dir(tempdir.path());
 
     snapbox.assert().stdout_matches(indoc! {r"
-        command: account add
+        command: account import
         add_profile: --add-profile flag was not set. No profile added to snfoundry.toml
     "});
 
@@ -316,7 +311,7 @@ pub async fn test_detect_deployed() {
         json!(
             {
                 "alpha-sepolia": {
-                  "my_account_add": {
+                  "my_account_import": {
                     "address": DEVNET_PREDEPLOYED_ACCOUNT_ADDRESS,
                     "class_hash": &DEVNET_OZ_CLASS_HASH_CAIRO_1.into_hex_string(),
                     "deployed": true,
@@ -332,39 +327,15 @@ pub async fn test_detect_deployed() {
 }
 
 #[tokio::test]
-pub async fn test_invalid_public_key() {
+pub async fn test_missing_arguments() {
     let args = vec![
         "account",
-        "add",
+        "import",
         "--url",
         URL,
         "--name",
-        "my_account_add",
-        "--address",
-        "0x123",
-        "--private-key",
-        "0x456",
-        "--public-key",
-        "0x457",
-        "--type",
-        "oz",
+        "my_account_import",
     ];
-
-    let snapbox = runner(&args);
-    let output = snapbox.assert().success();
-
-    assert_stderr_contains(
-        output,
-        indoc! {r"
-        command: account add
-        error: The private key does not match the public key
-        "},
-    );
-}
-
-#[tokio::test]
-pub async fn test_missing_arguments() {
-    let args = vec!["account", "add", "--url", URL, "--name", "my_account_add"];
 
     let snapbox = runner(&args);
     let output = snapbox.assert().failure();
@@ -375,7 +346,6 @@ pub async fn test_missing_arguments() {
         error: the following required arguments were not provided:
           --address <ADDRESS>
           --type <ACCOUNT_TYPE>
-          <--private-key <PRIVATE_KEY>|--private-key-file <PRIVATE_KEY_FILE_PATH>>
         "},
     );
 }
@@ -392,11 +362,11 @@ pub async fn test_private_key_from_file() {
         "--accounts-file",
         accounts_file,
         "account",
-        "add",
+        "import",
         "--url",
         URL,
         "--name",
-        "my_account_add",
+        "my_account_import",
         "--address",
         "0x123",
         "--private-key-file",
@@ -410,7 +380,7 @@ pub async fn test_private_key_from_file() {
     let snapbox = runner(&args).current_dir(temp_dir.path());
 
     snapbox.assert().stdout_matches(indoc! {r"
-        command: account add
+        command: account import
         add_profile: --add-profile flag was not set. No profile added to snfoundry.toml
     "});
 
@@ -422,7 +392,7 @@ pub async fn test_private_key_from_file() {
         json!(
             {
                 "alpha-sepolia": {
-                  "my_account_add": {
+                  "my_account_import": {
                     "address": "0x123",
                     "deployed": false,
                     "legacy": true,
@@ -441,9 +411,9 @@ pub async fn test_private_key_from_file() {
 pub async fn test_accept_only_one_private_key() {
     let args = vec![
         "account",
-        "add",
+        "import",
         "--name",
-        "my_account_add",
+        "my_account_import",
         "--address",
         "0x123",
         "--private-key",
@@ -465,11 +435,11 @@ pub async fn test_accept_only_one_private_key() {
 pub async fn test_invalid_private_key_file_path() {
     let args = vec![
         "account",
-        "add",
+        "import",
         "--url",
         URL,
         "--name",
-        "my_account_add",
+        "my_account_import",
         "--address",
         "0x123",
         "--private-key-file",
@@ -484,7 +454,7 @@ pub async fn test_invalid_private_key_file_path() {
     assert_stderr_contains(
         output,
         indoc! {r"
-        command: account add
+        command: account import
         error: Failed to obtain private key from the file my_private_key: No such file or directory (os error 2)
         "},
     );
@@ -505,11 +475,11 @@ pub async fn test_invalid_private_key_in_file() {
         "--accounts-file",
         "accounts.json",
         "account",
-        "add",
+        "import",
         "--url",
         URL,
         "--name",
-        "my_account_add",
+        "my_account_import",
         "--address",
         "0x123",
         "--private-key-file",
@@ -524,7 +494,7 @@ pub async fn test_invalid_private_key_in_file() {
     assert_stderr_contains(
         output,
         indoc! {r"
-        command: account add
+        command: account import
         error: Failed to obtain private key from the file my_private_key: Failed to create Felt from string
         "},
     );
@@ -542,11 +512,11 @@ pub async fn test_private_key_as_int_in_file() {
         "--accounts-file",
         accounts_file,
         "account",
-        "add",
+        "import",
         "--url",
         URL,
         "--name",
-        "my_account_add",
+        "my_account_import",
         "--address",
         DEVNET_PREDEPLOYED_ACCOUNT_ADDRESS,
         "--private-key-file",
@@ -568,7 +538,7 @@ pub async fn test_private_key_as_int_in_file() {
         json!(
             {
                 "alpha-sepolia": {
-                  "my_account_add": {
+                  "my_account_import": {
                     "address": DEVNET_PREDEPLOYED_ACCOUNT_ADDRESS,
                     "deployed": true,
                     "legacy": false,
@@ -593,11 +563,11 @@ pub async fn test_empty_config_add_profile() {
         "--accounts-file",
         accounts_file,
         "account",
-        "add",
+        "import",
         "--url",
         URL,
         "--name",
-        "my_account_add",
+        "my_account_import",
         "--address",
         DEVNET_PREDEPLOYED_ACCOUNT_ADDRESS,
         "--private-key",
@@ -611,7 +581,7 @@ pub async fn test_empty_config_add_profile() {
     let snapbox = runner(&args).current_dir(tempdir.path());
 
     snapbox.assert().stdout_matches(indoc! {r"
-        command: account add
+        command: account import
         add_profile: Profile random successfully added to snfoundry.toml
     "});
     let current_dir_utf8 = Utf8PathBuf::try_from(tempdir.path().to_path_buf()).unwrap();
@@ -619,6 +589,97 @@ pub async fn test_empty_config_add_profile() {
     let contents = fs::read_to_string(current_dir_utf8.join("snfoundry.toml"))
         .expect("Unable to read snfoundry.toml");
     assert!(contents.contains("[sncast.random]"));
-    assert!(contents.contains("account = \"my_account_add\""));
+    assert!(contents.contains("account = \"my_account_import\""));
     assert!(contents.contains(&format!("url = \"{URL}\"")));
+}
+
+#[tokio::test]
+pub async fn test_happy_case_valid_address_computation() {
+    let tempdir = tempdir().expect("Unable to create a temporary directory");
+    let accounts_file = "accounts.json";
+
+    let args = vec![
+        "--accounts-file",
+        accounts_file,
+        "account",
+        "import",
+        "--url",
+        URL,
+        "--name",
+        "my_account_import",
+        "--address",
+        "0x721c21e0cc9d37aec8e176797effd1be222aff6db25f068040adefabb7cfb6d",
+        "--private-key",
+        "0x2",
+        "--salt",
+        "0x3",
+        "--class-hash",
+        DEVNET_OZ_CLASS_HASH_CAIRO_0,
+        "--type",
+        "oz",
+    ];
+
+    let snapbox = runner(&args).current_dir(tempdir.path());
+
+    snapbox.assert().stdout_matches(indoc! {r"
+        command: account import
+        add_profile: --add-profile flag was not set. No profile added to snfoundry.toml
+    "});
+
+    let contents = fs::read_to_string(tempdir.path().join(accounts_file))
+        .expect("Unable to read created file");
+    let contents_json: serde_json::Value = serde_json::from_str(&contents).unwrap();
+    assert_eq!(
+        contents_json,
+        json!(
+            {
+                "alpha-sepolia": {
+                  "my_account_import": {
+                    "address": "0x721c21e0cc9d37aec8e176797effd1be222aff6db25f068040adefabb7cfb6d",
+                    "class_hash": DEVNET_OZ_CLASS_HASH_CAIRO_0,
+                    "deployed": false,
+                    "salt": "0x3",
+                    "legacy": true,
+                    "private_key": "0x2",
+                    "public_key": "0x759ca09377679ecd535a81e83039658bf40959283187c654c5416f439403cf5",
+                    "type": "open_zeppelin"
+                  }
+                }
+            }
+        )
+    );
+}
+
+#[tokio::test]
+pub async fn test_invalid_address_computation() {
+    let tempdir = tempdir().expect("Unable to create a temporary directory");
+    let accounts_file = "accounts.json";
+
+    let args = vec![
+        "--accounts-file",
+        accounts_file,
+        "account",
+        "import",
+        "--url",
+        URL,
+        "--name",
+        "my_account_import",
+        "--address",
+        "0x123",
+        "--private-key",
+        "0x456",
+        "--salt",
+        "0x789",
+        "--class-hash",
+        DEVNET_OZ_CLASS_HASH_CAIRO_0,
+        "--type",
+        "oz",
+    ];
+
+    let snapbox = runner(&args).current_dir(tempdir.path());
+    let computed_address = "0xaf550326d32c8106ef08d25cbc0dba06e5cd10a2201c2e9bc5ad4cbbce45e6";
+    snapbox.assert().stderr_matches(formatdoc! {r"
+        command: account import
+        error: Computed address {computed_address} does not match the provided address 0x123. Please ensure that the provided salt, class hash, and account type are correct.
+    "});
 }